Fixes:    Integer overflow in QSize::scale(). Task:     172712 AutoTest: included Details:  Don't convert to 32-bit too early.
[git-p4: depot-paths = "//depot/qt/main/": change = 272088]

+++ src/corelib/tools/qsize.cpp
@@ -186,7 +186,7 @@ void QSize::scale(const QSize &s, Qt::AspectRatioMode mode)
         ht = s.ht;
     } else {
         bool useHeight;
-        int rw = qint32(qint64(s.ht) * qint64(wd) / qint64(ht));
+        qint64 rw = qint64(s.ht) * qint64(wd) / qint64(ht);
 
         if (mode == Qt::KeepAspectRatio) {
             useHeight = (rw <= s.wd);
